backend/src/routes/ai_generator.py
import os
import json
from google import genai
from typing import Dict, Any
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

def generate_challenge_with_ai(difficulty: str) -> Dict[str, Any]:
    system_prompt = """You are an expert coding challenge creator. 
    Your task is to generate a coding question with multiple choice answers.
    The question should be appropriate for the specified difficulty level.

    For easy questions: Focus on basic syntax, simple operations, or common programming concepts.
    For medium questions: Cover intermediate concepts like data structures, algorithms, or language features.
    For hard questions: Include advanced topics, design patterns, optimization techniques, or complex algorithms.

    Return the challenge in the following JSON structure:
    {
        "title": "The question title",
        "options": ["Option 1", "Option 2", "Option 3", "Option 4"],
        "correct_answer_id": 0, // Index of the correct answer (0-3)
        "explanation": "Detailed explanation of why the correct answer is right"
    }

    Make sure the options are plausible but with only one clearly correct answer.
    """
    try:
        response = client.models.generate_content(
            model='gemini-2.5-flash',
            contents=f"{system_prompt}\n\nGenerate a {difficulty} difficulty coding challenge.",
            config={
                'response_mime_type': 'application/json',
                'temperature': 0.7
            }
        )

        challenge_data = json.loads(response.text)

        required_fields = ["title", "options", "correct_answer_id", "explanation"]
        for field in required_fields:
            if field not in challenge_data:
                raise ValueError(f"Missing required field: {field}")
            
        return challenge_data

    except Exception as e:
        print(f"GenAI error: {e}")
        return {
            "title": "Python List Appending",
            "options": ["append()", "add()", "push()", "insert()"],
            "correct_answer_id": 0,
            "explanation": "In Python, the append() method adds an item to the end of a list."
        }